test: move tck cypher-filtering tests into separate files
